Rewrite BasicURLNormalizer so URLs don't get double encoded
URL encoding is so tricky! And there are basically no good library
that handles it properly, mostly because of the lack of proper support
via the URL and URI classes.

In any event, I rewrote filterQueryElements so that it no longer
double encode the URLs. This is done by simply only modifying the
query string and leaving the other elements unchanged.

I added some 2 unit tests to validate my code.

Fixes #118