Fixed proxy auth bug in OkHTTP protocol & Enabled proxy file access via resources dirctory (#898)
* Fixed default proxy manager bug for non-proxy runs

Signed-off-by: Sam Ulrich <samson98ulrich@gmail.com>

* Added Metadata argument to getProxy function of ProxyManager

This argument is currently unused in the default implementations of ProxyManager but will enable custom implementations to have context as to what URL and environment the proxy will be used for. This will give valuable context for building much more complex proxy management logic.

Signed-off-by: Sam Ulrich <samson98ulrich@gmail.com>

* Added resource access for proxy files

MultiProxyManager will search the jar's inner resource directory for the desired proxy file. If found, it will load the proxy file from the resource directory making it easier to bundle the proxy file into the build across all nodes.

* Fixed bug where proxy was not used in OkHTTP is not auth was used

Typo caused by IDE auto fill cause the username field of the proxy to be used as the conditional for when a proxy was added to the request. This should have been triggered by the address of the proxy.